+++ tests/test_store.py
@@ -1,17 +1,17 @@
 import json
 import logging
+import mock
 import unittest
 
-import mock
-from structlog import wrap_logger
 from sqlalchemy.exc import IntegrityError, SQLAlchemyError
+from structlog import wrap_logger
 import testing.postgresql
 
-from tests.test_data import invalid_message, test_message, updated_message, missing_tx_id_message
-from tests.test_data import test_feedback_message, invalid_feedback_message
-
+import exporter
 import server
 from server import db, InvalidUsageError, logger
+from tests.test_data import invalid_message, test_message, updated_message, missing_tx_id_message
+from tests.test_data import test_feedback_message, invalid_feedback_message
 
 
 @testing.postgresql.skipIfNotInstalled
@@ -21,7 +21,20 @@ class TestStoreService(unittest.TestCase):
         'invalid': '/invalid_responses',
         'queue': '/queue',
         'healthcheck': '/healthcheck',
+        'comments': '/comments',
     }
+
+    rsi_data = {"data": {"11": "1/4/2016", "12": "31/10/2016", "20": "1800000", "21": "60000", "22": "705000", "23": "900",
+                         "24": "74", "25": "50", "26": "100", "27": "7400", "50": "205", "51": "84", "52": "10", "53": "73",
+                         "54": "24", "146": "Change comments included", "146a": "Yes",
+                         "146b": "In-store / online promotions", "146c": "Special events (e.g. sporting events)",
+                         "146d": "Calendar events (e.g. Christmas, Easter, Bank Holiday)", "146e": "Weather",
+                         "146f": "Store closures", "146g": "Store openings", "146h": "Other"},
+                "type": "uk.gov.ons.edc.eq:surveyresponse", "tx_id": "0d51ca67-98d9-4ae9-9187-2887f24c0a1f",
+                "origin": "uk.gov.ons.edc.eq", "version": "0.0.1",
+                "metadata": {"ru_ref": "12345678901A", "user_id": "789473423"},
+                "survey_id": "023", "collection": {"period": "1604", "exercise_sid": "hfjdskf", "instrument_id": "0215"},
+                "submitted_at": "2016-03-12T10:39:40Z"}
 
     logger = wrap_logger(logging.getLogger("TEST"))
 
@@ -114,7 +127,6 @@ class TestStoreService(unittest.TestCase):
         self.assertEqual(400, r.status_code)
 
     def test_get_responses_per_page(self):
-
         self.app.post(self.endpoints['responses'],
                       data=updated_message,
                       content_type='application/json')
@@ -158,3 +170,33 @@ class TestStoreService(unittest.TestCase):
             healthMock.side_effect = SQLAlchemyError
             r = self.app.get(self.endpoints['healthcheck'])
             self.assertEqual(500, r.status_code)
+
+    def test_create_workbook(self):
+        comments = [server.SurveyResponse('123', True, self.rsi_data)]
+        workbook, filename = exporter.create_comments_book('023', comments)
+        self.assertIs(type(workbook), bytes)
+        self.assertIs(type(filename), str)
+
+    def test_get_comments(self):
+        with mock.patch('server.get_all_comments_by_survey_id') as result_mock:
+            result_mock.return_value = [server.SurveyResponse('123', True, self.rsi_data)]
+
+            response = self.app.get(self.endpoints['comments'] + '/023')
+            self.assertEqual(200, response.status_code)
+            self.assertIs(type(response.get_data()), bytes)
+
+    def test_no_comments_for_survey(self):
+        with mock.patch('server.get_all_comments_by_survey_id') as result_mock:
+            result_mock.return_value = []
+
+            response = self.app.get(self.endpoints['comments'] + '/023')
+            self.assertEqual(200, response.status_code)
+            self.assertEqual(response.get_data(), b'{\n  "message": "No comments to export for 023"\n}\n')
+
+    def test_get_comments_has_errors(self):
+        with mock.patch('server.get_all_comments_by_survey_id') as result_mock:
+            result_mock.side_effect = SQLAlchemyError
+
+            response = self.app.get(self.endpoints['comments'] + '/023')
+            self.assertEqual(500, response.status_code)
+            self.assertEqual(response.get_data(), b'{\n  "message": "Database error", \n  "status": 500\n}\n')